+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioAdministrador.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Administrador;
 
-public interface RepositorioAdministrador extends JpaRepository<Administrador, Integer> {
+public interface RepositorioAdministrador extends JpaRepository<Administrador, Long> {
 
 }
