+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioAutonomia.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Autonomia;
 
-public interface RepositorioAutonomia extends JpaRepository<Autonomia, Integer> {
+public interface RepositorioAutonomia extends JpaRepository<Autonomia, Long> {
 
 }
