+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioComentario.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Comentario;
 
-public interface RepositorioComentario extends JpaRepository<Comentario, Integer> {
+public interface RepositorioComentario extends JpaRepository<Comentario, Long> {
 
 }
