+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioLista.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Lista;
 
-public interface RepositorioLista extends JpaRepository<Lista, Integer> {
+public interface RepositorioLista extends JpaRepository<Lista, Long> {
 
-	public List<Lista> findAllByResidenteId(Integer idResidente);
+	public List<Lista> findAllByResidenteId(Long idResidente);
 	
 }
