+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioLocalidad.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Localidad;
 
-public interface RepositorioLocalidad extends JpaRepository<Localidad, Integer> {
+public interface RepositorioLocalidad extends JpaRepository<Localidad, Long> {
 
 	public List<Localidad> findAllByProvinciaId(Integer idProvincia);
 }
