+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioMensaje.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Mensaje;
 
-public interface RepositorioMensaje extends JpaRepository<Mensaje, Integer> {
+public interface RepositorioMensaje extends JpaRepository<Mensaje, Long> {
 
-	public List<Mensaje> findAllByEmisorIdAndReceptorIdOrEmisorIdAndReceptorId(Integer idEmisor, Integer idReceptor,
-			Integer idReceptor2, Integer idEmisor2);
+	public List<Mensaje> findAllByEmisorIdAndReceptorIdOrEmisorIdAndReceptorId(Long idEmisor, Long idReceptor,
+			Long idReceptor2, Long idEmisor2);
 
-	public List<Mensaje> findAllByEmisorId(Integer idEmisor);
+	public List<Mensaje> findAllByEmisorId(Long idEmisor);
 
-	public List<Mensaje> findAllByReceptorId(Integer idReceptor);
+	public List<Mensaje> findAllByReceptorId(Long idReceptor);
 }
