+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioProvincia.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Provincia;
 
-public interface RepositorioProvincia extends JpaRepository<Provincia, Integer> {
+public interface RepositorioProvincia extends JpaRepository<Provincia, Long> {
 
 	public List<Provincia> findAllByAutonomiaId(Integer autonomiaId);
 }
