+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioTipoGuardiasFindesFestivos.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.TipoGuardiasFindesFestivos;
 
-public interface RepositorioTipoGuardiasFindesFestivos extends JpaRepository<TipoGuardiasFindesFestivos, Integer> {
+public interface RepositorioTipoGuardiasFindesFestivos extends JpaRepository<TipoGuardiasFindesFestivos, Long> {
 
 }
