+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioTipoTrabajoFinResidencia.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.TipoTrabajoFinResidencia;
 
-public interface RepositorioTipoTrabajoFinResidencia extends JpaRepository<TipoTrabajoFinResidencia, Integer> {
+public interface RepositorioTipoTrabajoFinResidencia extends JpaRepository<TipoTrabajoFinResidencia, Long> {
 
 }
