+++ src/main/java/com/mifse/backend/persistencia/repositorios/RepositorioTitulacion.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mifse.backend.persistencia.modelos.Titulacion;
 
-public interface RepositorioTitulacion extends JpaRepository<Titulacion, Integer> {
+public interface RepositorioTitulacion extends JpaRepository<Titulacion, Long> {
 
 }
