+++ src/main/java/com/mifse/backend/servicios/ServicioEmail.java
@@ -2,7 +2,7 @@ package com.mifse.backend.servicios;
 
 public interface ServicioEmail {
 
-	public void enviarEmailVerificacion(String nombreDestinatario, String emailDestinatario, Integer idUsuario);
+	public void enviarEmailVerificacion(String nombreDestinatario, String emailDestinatario, Long idUsuario);
 	
 	public void enviarEmailContacto(String remitente, String asunto, String cuerpo);
 }
