+++ src/main/java/com/mifse/backend/servicios/ServicioMensaje.java
@@ -8,11 +8,11 @@ import com.mifse.backend.persistencia.modelos.Usuario;
 
 public interface ServicioMensaje {
 
-	public List<Mensaje> obtenerTodosPorIdEmisorYIdReceptor(Integer idEmisor, Integer idReceptor);
+	public List<Mensaje> obtenerTodosPorIdEmisorYIdReceptor(Long idEmisor, Long idReceptor);
 
-	public Set<Usuario> obtenerUsuariosConMensajesIntercambiados(Integer idUsuario);
+	public Set<Usuario> obtenerUsuariosConMensajesIntercambiados(Long idUsuario);
 
-	public Set<Usuario> obtenerUsuariosConMensajesIntercambiadosSinLeer(Integer idUsuario);
+	public Set<Usuario> obtenerUsuariosConMensajesIntercambiadosSinLeer(Long idUsuario);
 
 	public Mensaje guardar(Mensaje mensaje);
